archive/monitoring_tool/requirements.txt
aiohttp==3.7.4.post0
aiosignal==1.3.1
async-timeout==3.0.1
attrs==21.4.0
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==2.1.1
colorama==0.4.6
docx==0.2.4
frozenlist==1.3.3
humanfriendly==10.0
idna==3.3
lxml==4.9.2
multidict==6.0.2
numpy==1.24.2
Pillow==9.4.0
plyer==2.1.0
psutil==5.9.4
pyfiglet==0.8.post1
pygame==2.1.3
pyperclip==1.8.2
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
pyreadline3==3.4.1
pySMART==1.2.3
pystray==0.19.4
python-dotenv==0.21.1
requests==2.28.2
six==1.16.0
typing_extensions==4.2.0
urllib3==1.26.14
wxPython==4.2.0
yarl==1.7.2
